Allow to renumerate generated graph
Renumerating a graph will change all vertex IDs so that all edges go from a vertex with a lower ID to a vertex with a higher ID. All vertices with no incoming edges are also guaranteed to have the lowest numbers. This change is mainly for aesthetic purposes and can be omitted for faster graph generation.